chore(switch): rename toggle to switch
