+++ .codesandbox/ci.json
@@ -62,6 +62,7 @@
     "/packages/paste-core/layout/stack",
     "/packages/paste-style-props",
     "/packages/paste-libraries/styling",
+    "/packages/paste-core/components/switch",
     "/packages/paste-core/components/table",
     "/packages/paste-core/components/tabs",
     "/packages/paste-core/primitives/tabs",
@@ -70,7 +71,6 @@
     "/packages/paste-theme",
     "/packages/paste-core/components/time-picker",
     "/packages/paste-core/components/toast",
-    "/packages/paste-core/components/toggle",
     "/packages/paste-core/components/tooltip",
     "/packages/paste-core/primitives/tooltip",
     "/packages/paste-core/components/truncate",
