+++ packages/paste-codemods/tools/.cache/mappings.json
@@ -122,6 +122,7 @@
   "Separator": "@twilio-paste/core/separator",
   "SkeletonLoader": "@twilio-paste/core/skeleton-loader",
   "Spinner": "@twilio-paste/core/spinner",
+  "Switch": "@twilio-paste/core/switch",
   "TBody": "@twilio-paste/core/table",
   "TBodyPropTypes": "@twilio-paste/core/table",
   "TFoot": "@twilio-paste/core/table",
@@ -151,7 +152,6 @@
   "ToastContainer": "@twilio-paste/core/toast",
   "Toaster": "@twilio-paste/core/toast",
   "useToaster": "@twilio-paste/core/toast",
-  "Toggle": "@twilio-paste/core/toggle",
   "Tooltip": "@twilio-paste/core/tooltip",
   "useTooltipState": "@twilio-paste/core/tooltip",
   "Truncate": "@twilio-paste/core/truncate",
