+++ packages/paste-core/components/switch/__tests__/index.spec.tsx
