+++ packages/paste-core/components/switch/build.js
