+++ packages/paste-core/components/switch/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@twilio-paste/toggle",
+  "name": "@twilio-paste/switch",
   "version": "0.0.0",
   "category": "interaction",
   "status": "production",
-  "description": "A toggle is an interactive binary switch.",
+  "description": "A switch is an interactive binary control.",
   "author": "Twilio Inc.",
   "license": "MIT",
   "main:dev": "src/index.tsx",
