+++ packages/paste-core/components/switch/src/SwitchHelpText.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import type {ToggleHelpTextProps} from './types';
+import type {SwitchHelpTextProps} from './types';
 import {HelpText} from '@twilio-paste/help-text';
 import {Box, safelySpreadBoxProps} from '@twilio-paste/box';
 
-const ToggleHelpText = React.forwardRef<HTMLSpanElement, ToggleHelpTextProps>(
-  ({element = 'TOGGLE_HELP_TEXT', children, helpTextId, ...props}, ref) => {
+const SwitchHelpText = React.forwardRef<HTMLSpanElement, SwitchHelpTextProps>(
+  ({element = 'SWITCH_HELP_TEXT_WRAPPER', children, helpTextId, ...props}, ref) => {
     return (
       // @ts-ignore TODO:no 50px tokens
       <Box {...safelySpreadBoxProps(props)} element={element} marginLeft="50px" ref={ref}>
@@ -17,12 +17,12 @@ const ToggleHelpText = React.forwardRef<HTMLSpanElement, ToggleHelpTextProps>(
   }
 );
 
-ToggleHelpText.displayName = 'ToggleHelpText';
+SwitchHelpText.displayName = 'SwitchHelpText';
 
-ToggleHelpText.propTypes = {
+SwitchHelpText.propTypes = {
   children: PropTypes.node.isRequired,
   helpTextId: PropTypes.string.isRequired,
   element: PropTypes.string,
 };
 
-export {ToggleHelpText};
+export {SwitchHelpText};
