+++ packages/paste-core/components/switch/src/constants.ts
