+++ packages/paste-core/components/switch/src/index.tsx
@@ -1,0 +1,2 @@
+export * from './SwitchContainer';
+export * from './Switch';
