+++ packages/paste-core/components/switch/tsconfig.json
