+++ packages/paste-core/core-bundle/src/index.tsx
@@ -45,6 +45,7 @@ export * from '@twilio-paste/sibling-box';
 export * from '@twilio-paste/skeleton-loader';
 export * from '@twilio-paste/spinner';
 export * from '@twilio-paste/stack';
+export * from '@twilio-paste/switch';
 export * from '@twilio-paste/table';
 export * from '@twilio-paste/tabs';
 export * from '@twilio-paste/tabs-primitive';
@@ -52,7 +53,6 @@ export * from '@twilio-paste/text';
 export * from '@twilio-paste/textarea';
 export * from '@twilio-paste/time-picker';
 export * from '@twilio-paste/toast';
-export * from '@twilio-paste/toggle';
 export * from '@twilio-paste/tooltip';
 export * from '@twilio-paste/tooltip-primitive';
 export * from '@twilio-paste/truncate';
