+++ packages/paste-core/core-bundle/src/switch.tsx
@@ -1,0 +1,1 @@
+export * from '@twilio-paste/switch';
