--- packages/paste-core/core-bundle/src/toggle.tsx
@@ -1,1 +1,0 @@
-export * from '@twilio-paste/toggle';
